# docs/swagger.yaml
definitions:
...
  req.ForgotPasswordReq:
    properties:
      phone:
        type: string
    required:
    - phone
    type: object
  req.NewAddressReq:
    properties:
      district:
        type: string
      houseName:
        type: string
      name:
        type: string
      phone:
        type: string
      pinCode:
        type: string
      state:
        type: string
      street:
        type: string
    required:
    - district
    - houseName
    - name
    - phone
    - pinCode
    - state
    - street
    type: object
...
  req.UserLoginReq:
    properties:
      email:
        type: string
      password:
        minLength: 3
        type: string
    required:
    - email
    - password
    type: object
...
  res.CommonRes:
    properties:
      error:
        type: string
      message:
        type: string
      result: {}
      status:
        type: string
    type: object
...
  res.UserDetailsRes:
    properties:
      message:
        type: string
      status:
        type: string
      user:
        $ref: '#/definitions/entities.User'
    type: object
...
  res.UserLoginRes:
    properties:
      message:
        type: string
      status:
        type: string
      token:
        type: string
    type: object
  res.ViewAddressListRes:
    properties:
      addressList:
        items:
          $ref: '#/definitions/entities.Address'
        type: array
      message:
        type: string
      status:
        type: string
    type: object
  res.ViewAddressRes:
    properties:
      address:
        $ref: '#/definitions/entities.Address'
      message:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /forgotPassword:
    post:
      consumes:
      - application/json
      description: Forgot user password
      parameters:
      - description: Forgot password request
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/req.ForgotPasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully sent otp to number related to the given email
          schema:
            $ref: '#/definitions/res.CommonRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.CommonRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.CommonRes'
      summary: Forgot Password
      tags:
      - User
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate and log in as a user
      parameters:
      - description: User login request
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/req.UserLoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully logged in
          schema:
            $ref: '#/definitions/res.UserLoginRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.CommonRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.CommonRes'
      summary: User Login
      tags:
      - User
...
  /profile:
    get:
      consumes:
      - application/json
      description: Get user details
      produces:
      - application/json
      responses:
        "200":
          description: Successfully fetched user details
          schema:
            $ref: '#/definitions/res.UserDetailsRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.CommonRes'
        "401":
          description: Unauthorized Access
          schema:
            $ref: '#/definitions/res.CommonRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.CommonRes'
      security:
      - Bearer: []
      summary: View User Details
      tags:
      - User Profile
  /profile/addAddress:
    post:
      consumes:
      - application/json
      description: Add a new address for the user
      parameters:
      - description: New address request
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/req.NewAddressReq'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully added address
          schema:
            $ref: '#/definitions/res.CommonRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.CommonRes'
        "401":
          description: Unauthorized Access
          schema:
            $ref: '#/definitions/res.CommonRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.CommonRes'
      security:
      - Bearer: []
      summary: Add an address
      tags:
      - User Profile
  /profile/address:
    get:
      consumes:
      - application/json
      description: View details of all addresses for the user
      produces:
      - application/json
      responses:
        "200":
          description: Successfully fetched addresses
          schema:
            $ref: '#/definitions/res.ViewAddressListRes'
        "401":
          description: Unauthorized Access
          schema:
            $ref: '#/definitions/res.CommonRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.CommonRes'
      security:
      - Bearer: []
      summary: View all addresses
      tags:
      - User Profile
  /profile/address/{id}:
    get:
      consumes:
      - application/json
      description: View details of a specific address for the user
      parameters:
      - description: Address ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully fetched address
          schema:
            $ref: '#/definitions/res.ViewAddressRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.CommonRes'
        "401":
          description: Unauthorized Access
          schema:
            $ref: '#/definitions/res.CommonRes'
        "404":
          description: Address not found
          schema:
            $ref: '#/definitions/res.CommonRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.CommonRes'
      security:
      - Bearer: []
      summary: View a specific address
      tags:
      - User Profile
    put:
      consumes:
      - application/json
      description: Update existing address for the user
      parameters:
      - description: Update address request
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/req.UpdateAddressReq'
      - description: Address ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated address
          schema:
            $ref: '#/definitions/res.CommonRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.CommonRes'
        "401":
          description: Unauthorized Access
          schema:
            $ref: '#/definitions/res.CommonRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.CommonRes'
      security:
      - Bearer: []
      summary: Update address
      tags:
      - User Profile
  /profile/changePassword:
    post:
      consumes:
      - application/json
      description: Change existing password of user
      parameters:
      - description: Change user password request
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/req.ChangePasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully changed user password details
          schema:
            $ref: '#/definitions/res.CommonRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.CommonRes'
        "401":
          description: Unauthorized Access
          schema:
            $ref: '#/definitions/res.CommonRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.CommonRes'
      security:
      - Bearer: []
      summary: Change Password
      tags:
      - User Profile
  /profile/edit:
    patch:
      consumes:
      - application/json
      description: Update user name, email details
      parameters:
      - description: Edit user details request
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/req.UpdateUserDetailsReq'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated user details
          schema:
            $ref: '#/definitions/res.UserDetailsRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.CommonRes'
        "401":
          description: Unauthorized Access
          schema:
            $ref: '#/definitions/res.CommonRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.CommonRes'
      security:
      - Bearer: []
      summary: Update User Details
      tags:
      - User Profile
  /resetPassword:
    post:
      consumes:
      - application/json
      description: Reset user password
      parameters:
      - description: Reset password request
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/req.ResetPasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: Reset Password Successfull
          schema:
            $ref: '#/definitions/res.CommonRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.CommonRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.CommonRes'
      summary: Reset Password
      tags:
      - User
...
  /search/sellers:
    get:
      consumes:
      - application/json
      description: Search for sellers based on a query
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully fetched sellers
          schema:
            $ref: '#/definitions/res.SellerListRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.CommonRes'
      summary: Search sellers
      tags:
      - Common
...
  /sendOtp:
    post:
      consumes:
      - application/json
      description: Send OTP to the user's registered phone number
      produces:
      - application/json
      responses:
        "200":
          description: Successfully sent OTP
          schema:
            $ref: '#/definitions/res.CommonRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.CommonRes'
        "401":
          description: Unauthorized Access
          schema:
            $ref: '#/definitions/res.CommonRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.CommonRes'
      security:
      - Bearer: []
      summary: Send OTP
      tags:
      - User
  /signup:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: User sign-up request
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/req.UserSignUpReq'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully signed up
          schema:
            $ref: '#/definitions/res.UserLoginRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.CommonRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.CommonRes'
      summary: Sign up as a user
      tags:
      - User
  /user/sellers:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of sellers for the user
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: p
        type: string
      - description: Number of items per page
        in: query
        name: l
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully fetched sellers
          schema:
            $ref: '#/definitions/res.SellerListRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.CommonRes'
      summary: Get paginated list of sellers
      tags:
      - Common
  /user/sellers/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific seller by ID for the user
      parameters:
      - description: Seller ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully fetched seller
          schema:
            $ref: '#/definitions/res.SingleSellerRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.CommonRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.CommonRes'
      summary: Get a seller
      tags:
      - Common
  /verifyOtp:
    post:
      consumes:
      - application/json
      description: Verify OTP for the user's registered phone number
      parameters:
      - description: User OTP verification request
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/req.UserVerifyOtpReq'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully verified OTP
          schema:
            $ref: '#/definitions/res.CommonRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.CommonRes'
        "401":
          description: Unauthorized Access
          schema:
            $ref: '#/definitions/res.CommonRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.CommonRes'
      security:
      - Bearer: []
      summary: Verify OTP
      tags:
      - User
securityDefinitions:
  Bearer:
    description: 'Authentication using a JSON Web Token (JWT). The token should be
      included in the request header named "Authorization". The format of the header
      is: Authorization: Bearer <token>. Replace `<token>` with the actual JWT token.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"